requirements.txt
fastapi~=0.115.6
sqlalchemy==2.0.30
sqlalchemy-utils==0.41.2
uvicorn[standard]~=0.18.2
python-dotenv~=1.0.1
bcrypt~=4.0.1
orjson==3.10.15
shortuuid~=1.0.12
pydantic[email]
pyjwt==2.10.1
psycopg2-binary==2.9.10